debugging optical_flow.cpp itegration in aknn.cpp

+++ pipeline/optical_flow.h
@@ -2,4 +2,4 @@
 
 using namespace cv;
 using namespace std;
-vector<Point2f> get_offset(Point2f point, int current_frame, int next_frame_offset);
+vector<Point2f> get_offset(Point2f point, short current_frame, short next_frame_offset);
